requirements.txt
streamlit
yt-dlp
faster-whisper
torch
langchain
langchain-community
chromadb
groq
sentence_transformers
faiss_cpu
numpy
python-dotenv
gtts
python-docx
reportlab
streamlit
yt-dlp
faster-whisper
sentence-transformers
faiss-cpu
groq
python-dotenv
better_profanity
spacy
transformers
reportlab
python-docx
pymongo
pyannote.audio